vine: tune parameter immediate-recovery (#3517)
* vine: tune parameter immediate-recovery

Submit recovery tasks for temporary files as soon as their worker is
lost.

Use case: topEFT sorts tasks according to how large they are when
constructing accumulation tasks. This means that it takes a long time
to discover that a result that was expensive to compute was lost. With
this change, recovery tasks are created as soon as the worker is gone.
Since all tmp files are used as inputs at least once, and topEFT removes
tmp files as soon as they are used, this change saves time at the
workflow rampdown.

* format